fix: reschedule job twice after updating
